+++ snapshot-tests-directory-params/src/main/java/de/skuzzle/test/snapshots/directoryparams/TestFile.java
@@ -12,6 +12,9 @@ import java.util.Set;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.apiguardian.api.API;
+import org.apiguardian.api.API.Status;
+
 /**
  * Pointer to a file that has been listed in a directory by {@link FilesFrom}. Provides a
  * few methods for conveniently working with that file.
@@ -19,6 +22,7 @@ import java.util.regex.Pattern;
  * @see FilesFrom
  * @author Simon Taddiken
  */
+@API(status = Status.EXPERIMENTAL)
 public final class TestFile {
 
     private final Path file;
@@ -27,14 +31,29 @@ public final class TestFile {
         this.file = Objects.requireNonNull(file, "file must not be null");
     }
 
+    /**
+     * This file as Path.
+     *
+     * @return
+     */
     public Path file() {
         return this.file;
     }
 
+    /**
+     * The directory that contains this file.
+     *
+     * @return This file's directory.
+     */
     public Path directory() {
         return this.file.getParent();
     }
 
+    /**
+     * The name of this file without extensions
+     *
+     * @return Name without extension.
+     */
     public String name() {
         final String fileName = file.getFileName().toString();
         final int dot = fileName.lastIndexOf('.');
@@ -43,18 +62,42 @@ public final class TestFile {
                 : fileName.substring(0, dot);
     }
 
+    /**
+     * The full name of this file including extension.
+     *
+     * @return The full name.
+     */
     public String nameWithExtension() {
         return file.getFileName().toString();
     }
 
+    /**
+     * Returns this file's extension without leading dot. If you file has no extension
+     * (that is, {@link #nameWithExtension()} does not contain a '.') an empty String is
+     * returned.
+     *
+     * @return This file's extension without leading dot.
+     */
     public String extension() {
         return FileExtension.withoutLeadingDot(file);
     }
 
+    /**
+     * Opens a new stream to read the file's content.
+     *
+     * @return The stream.
+     * @throws IOException If an IO error occurs.
+     */
     public InputStream openStream() throws IOException {
         return Files.newInputStream(file);
     }
 
+    /**
+     * Fully reads this file into a newly allocated byte array.
+     *
+     * @return The contents of the file as byte array.
+     * @throws IOException If an IO error occurs.
+     */
     public byte[] asBinary() throws IOException {
         try (var inputStream = openStream()) {
             return inputStream.readAllBytes();