+++ RELEASE_NOTES.md
@@ -1,8 +1,6 @@
 [![Maven Central](https://img.shields.io/static/v1?label=MavenCentral&message=0.0.7&color=blue)](https://search.maven.org/artifact/de.skuzzle.test/snapshot-tests-parent/0.0.7/jar) [![JavaDoc](https://img.shields.io/static/v1?label=JavaDoc&message=0.0.7&color=orange)](http://www.javadoc.io/doc/de.skuzzle.test/snapshot-tests-parent/0.0.7)
 
-* Allow to customize the Xml-Unit assertion via `XmlSnapshot.compareUsing(Consumer<CompareAssert>)`
-* Restructure modules
-* Experimental normalization API
+* Add ApiGuardian annotations to all public API
 
 Maven Central coordinates for this release: